Cleanup hass.io component
